Bugfix for convolution ops derivatives
The calculation of convolution and max pooling partial derivatives had
issues that have been corrected. Commit also fixes/improves other aspects
of our CNN model.

However, while testing the CIFAR example, results don't seem to improve
after training the model (at least not significantly). This might be an
issue with the way our example is implemented, with the CNN itself or
with derivatives computations. This should be fixed as soon as possible
in future commits to validate our CNN implementation.

+++ include/convolution.hpp
@@ -10,6 +10,7 @@
 
 #include <vector>
 #include <memory>
+#include <iostream>
 
 #include <Eigen/Dense>
 
@@ -90,7 +91,7 @@ private:
 	FlatteningNode(
 		std::vector<long> shape,
 		Eigen::Array<T, Eigen::Dynamic, Eigen::Dynamic> xVal, int xDep,
-		std::vector<unsigned> newSize
+		std::vector<long> newSize
 	);
 
 	Eigen::Array<T, Eigen::Dynamic, Eigen::Dynamic> incrementGradient(
@@ -98,7 +99,7 @@ private:
 			unsigned &j
 	);
 
-	std::vector<unsigned> size = {};
+	std::vector<long> size = {};
 
 	friend ts::Tensor<T> flattening<>(const ts::Tensor<T> &x);
 };
